Fixed Capturing bug
placing the piece back down in the old position would capture the piece. This fixes that.
Currently capturing the same type of piece (e.g. white pawn trying to capture another white pawn)  causes strange behaviour. This is not an issue, however, as in the future this move will not be possible to be made